Update version to 0.10.0.1.
This required rewriting the build, as it has changed considerably.  I
replicated the directory structure of the binary kafka distribution,
which pretty much just has the bin scripts and the jar files (in
'libs').  Put the log dir in /var/logs, but I haven't done
the configuration changes to make that work yet.  This builds
an RPM, but doubtless does not yet work if installed, since I
haven't tested doing so.