Init script fixups, more macros in spec.
Made the log dir an environment variable, and made the two init scripts
consistent in the way they handle variables vs the possible sysconfig
file (functionally the same, but might as well be consistent).
Converted some more hardcoding to macros in the spec file, and
introduced some new defines.  Added a clause to create the zookeper user
and group, and the zookeeper logging directory.  Moved the
cleanup of the build directory to %clean where it is in most
specs I've looked at.  Added some blank lines to make inter-section
spacing consistent.