Fix logging and data storage.  Working RPM.
Needed some more tweaks to get all the logging to go to /var/log, and
figured out how to move the data into /var/lib.  So now there is a
default config that allows the installed init scripts to start up
zookeeper and a single kafka broker running on the project default
ports.  Added chkconfig to enable it on boot, since if you are
installing it that's almost certainly what you want.  A production
installation would have to update the configs, of course.